api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
aiomysql==0.2.0
pymysql==1.1.1
alembic==1.14.0
pydantic-settings==2.6.1
email-validator==2.2.0
python-jose[cryptography]==3.3.0
argon2-cffi==23.1.0
passlib[argon2]==1.7.4
python-multipart==0.0.17
celery==5.4.0
redis==5.2.0
httpx==0.28.1
feedparser==6.0.11
beautifulsoup4==4.12.3
pypdf==5.1.0
numpy==2.1.3
pytest==8.3.4
pytest-asyncio==0.24.0
asgi-lifespan==2.1.0
aiosqlite==0.20.0
